# lib/api-zod/src/generated/types/servicesItem.ts
/**
 * Generated by orval v8.15.0 🍺
 * Do not edit manually.
 * Api
 * Architectural Portfolio API
 * OpenAPI spec version: 0.1.0
 */
import type { ProjectSummary } from './projectSummary';

export interface ServicesItem {
  id: number;
  name: string;
  slug: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  longDescription?: string | null;
  /** @nullable */
  imageUrl?: string | null;
  /** @nullable */
  galleryImages?: string | null;
  published: boolean;
  featured?: boolean;
  projectCount?: number;
  projects?: ProjectSummary[];
}